Address review feedback on IPv6/CIDR egress filtering
Signal-time cleanup now removes only what it created
----------------------------------------------------
`force_cleanup` runs on the watchdog thread and had no access to the
manager's `CreatedResources`, so it hardcoded `v4_chain: true` and
`v6_chain: true`.  On a signal that arrived before the chains existed it
issued `-F`/`-X` against names it did not own, and because chain names
truncate at 20 characters that name can belong to a different container.

`CreatedResources` now lives in `signal_cleanup::ActiveSandbox` alongside
the container name and veth, behind the same mutex, so the watchdog takes
one coherent snapshot and can never pair one container's identity with
another's ownership record.  Every creation site publishes incrementally,
so a signal arriving mid-apply still sees the half-built set instead of an
empty one.  `force_cleanup` takes the record as a parameter and returns
immediately when it is empty, running zero iptables commands.

`teardown_created` now returns the residual set — ownership bits are
cleared only when the removal command actually succeeds — so a failed
removal is not recorded as a completed one.

IPv4-mapped IPv6 destinations no longer fail open
-------------------------------------------------
A dual-stack socket sending to `::ffff:1.2.3.4` makes Linux emit a real
IPv4 packet, so an `ip6tables -d ::ffff:1.2.3.4` rule never matches and a
mapped `blockedHosts` entry silently failed open under an allow default
policy.  Mapped literals and mapped CIDRs are now rewritten to their IPv4
form and filed into the IPv4 bucket.  Prefixes shorter than /96 span
outside the mapped range and are deliberately left as IPv6.

An unreadable /proc is no longer a confirmed "IPv6 is off"
----------------------------------------------------------
`NotFound` on `/proc/net/if_inet6` now maps to `Inactive` only when
`/proc/net` itself exists.  In a mount namespace without `/proc` the
answer is `Unknown`, which fails closed.

Review housekeeping
-------------------
Configs added by this change now declare `0.8.0-alpha`, matching
`CURRENT_SCHEMA_VERSION`.  This selects an existing schema version; it
does not modify any schema.

The five `#[path]` spec-test files are folded into the inline `mod tests`
in `network_iptables.rs`, matching the repo convention (123 inline test
modules against 5 `#[path]` files, all five of which this change added).

The enforcement-mode test asserted the production predicate against a
second copy of the same predicate, so both could be wrong together.  It
now asserts literal expected values.

Copilot-Session: 52fb9e8c-4a48-435c-9964-8ec0fee653c6
