+++ tests/configs/lxc_network_cidr_boundary.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.6.0-alpha",
+  "version": "0.8.0-alpha",
   "containerId": "CLI-LXC-Network-CIDR-Boundary",
   "containment": "lxc",
   "process": {
