+++ tests/configs/lxc_network_dualstack_hostname.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.6.0-alpha",
+  "version": "0.8.0-alpha",
   "containerId": "CLI-LXC-Network-Dualstack-Hostname",
   "containment": "lxc",
   "process": {
