+++ tests/configs/lxc_network_invalid_cidr.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.6.0-alpha",
+  "version": "0.8.0-alpha",
   "containerId": "CLI-LXC-Network-Invalid-CIDR",
   "containment": "lxc",
   "process": {
